refactor: move worker outside of utils
