+++ crates/cli/src/utils/mod.rs
@@ -1,16 +1,17 @@
 mod args;
+mod worker;
 
 pub use args::{InputArgs, OutputArgs};
+pub use worker::{Items, PathWorker, StdInWorker, Worker};
 
 use crate::lang::SgLang;
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::{Context, Result};
 use crossterm::{
   event::{self, Event, KeyCode},
   execute,
   terminal::{self, EnterAlternateScreen, LeaveAlternateScreen},
 };
-use ignore::{DirEntry, WalkParallel, WalkState};
 
 use ast_grep_config::{CombinedScan, PreScan, RuleCollection};
 use ast_grep_core::Pattern;
@@ -22,7 +23,6 @@ use std::io::stdout;
 use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
-use std::sync::{mpsc, Arc};
 
 type AstGrep = ast_grep_core::AstGrep<StrDoc<SgLang>>;
 
@@ -76,128 +76,6 @@ pub fn prompt(prompt_text: &str, letters: &str, default: Option<char>) -> Result
     }
     eprintln!("Unrecognized command, try again?")
   }
-}
-
-/// A trait to abstract how ast-grep discovers work Items.
-///
-/// It follows multiple-producer-single-consumer pattern.
-/// ast-grep will produce items in one or more separate thread(s) and
-/// `consumer_items` in the main thread, blocking the function return.
-/// Worker at the moment has two main flavors:
-/// * PathWorker: discovers files on the file system, based on ignore
-/// * StdInWorker: parse text content from standard input stream
-pub trait Worker: Sync + Send {
-  /// The item to send between producer/consumer threads.
-  /// It is usually parsed tree-sitter Root with optional data.
-  type Item: Send + 'static;
-  /// `consume_items` will run in a separate single thread.
-  /// printing matches or error reporting can happen here.
-  fn consume_items(&self, items: Items<Self::Item>) -> Result<()>;
-}
-
-/// A trait to abstract how ast-grep discovers, parses and processes files.
-///
-/// It follows multiple-producer-single-consumer pattern.
-/// ast-grep discovers files in parallel by `build_walk`.
-/// Then every file is parsed and filtered in `produce_item`.
-/// Finally, `produce_item` will send `Item` to the consumer thread.
-pub trait PathWorker: Worker {
-  /// WalkParallel will determine what files will be processed.
-  fn build_walk(&self) -> Result<WalkParallel>;
-  /// Parse and find_match can be done in `produce_item`.
-  fn produce_item(&self, path: &Path) -> Option<Vec<Self::Item>>;
-
-  fn run_path(self) -> Result<()>
-  where
-    Self: Sized + 'static,
-  {
-    run_worker(Arc::new(self))
-  }
-}
-
-pub trait StdInWorker: Worker {
-  fn parse_stdin(&self, src: String) -> Option<Self::Item>;
-
-  fn run_std_in(&self) -> Result<()> {
-    let source = std::io::read_to_string(std::io::stdin())?;
-    if let Some(item) = self.parse_stdin(source) {
-      self.consume_items(Items::once(item)?)
-    } else {
-      Ok(())
-    }
-  }
-}
-
-pub struct Items<T>(mpsc::Receiver<T>);
-impl<T> Iterator for Items<T> {
-  type Item = T;
-  fn next(&mut self) -> Option<Self::Item> {
-    if let Ok(match_result) = self.0.recv() {
-      Some(match_result)
-    } else {
-      None
-    }
-  }
-}
-impl<T> Items<T> {
-  fn once(t: T) -> Result<Self> {
-    let (tx, rx) = mpsc::channel();
-    // use write to avoid send/sync trait bound
-    match tx.send(t) {
-      Ok(_) => (),
-      Err(e) => return Err(anyhow!(e.to_string())),
-    };
-    Ok(Items(rx))
-  }
-}
-
-fn filter_result(result: Result<DirEntry, ignore::Error>) -> Option<PathBuf> {
-  let entry = match result {
-    Ok(entry) => entry,
-    Err(err) => {
-      eprintln!("ERROR: {}", err);
-      return None;
-    }
-  };
-  if !entry.file_type()?.is_file() {
-    return None;
-  }
-  let path = entry.into_path();
-  // TODO: is it correct here? see https://github.com/ast-grep/ast-grep/issues/1343
-  match path.strip_prefix("./") {
-    Ok(p) => Some(p.to_path_buf()),
-    Err(_) => Some(path),
-  }
-}
-
-fn run_worker<W: PathWorker + ?Sized + 'static>(worker: Arc<W>) -> Result<()> {
-  let (tx, rx) = mpsc::channel();
-  let w = worker.clone();
-  let walker = worker.build_walk()?;
-  // walker run will block the thread
-  std::thread::spawn(move || {
-    let tx = tx;
-    walker.run(|| {
-      let tx = tx.clone();
-      let w = w.clone();
-      Box::new(move |result| {
-        let Some(p) = filter_result(result) else {
-          return WalkState::Continue;
-        };
-        let Some(items) = w.produce_item(&p) else {
-          return WalkState::Continue;
-        };
-        for result in items {
-          match tx.send(result) {
-            Ok(_) => continue,
-            Err(_) => return WalkState::Quit,
-          }
-        }
-        WalkState::Continue
-      })
-    });
-  });
-  worker.consume_items(Items(rx))
 }
 
 fn read_file(path: &Path) -> Option<String> {
